Rename all files to kebab-case
