+++ src/components/tournament/team-details-dialog.tsx
