+++ src/components/tournament/tournament-details.tsx
